requirements.txt
anytree==2.9.0
cmder==3.8
dill==0.3.7
loguru==0.7.2
meeko==0.5.0
multiprocess==0.70.15
numpy==1.26.0
pandarallel==1.6.5
pandas==2.1.0
parquet-cli==1.3
pathos==0.3.1
Pillow==10.0.1
pox==0.3.3
ppft==1.7.6.7
psutil==5.9.5
pyarrow==13.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
rdkit==2023.3.3
scipy==1.11.2
seqflow==0.1.4
six==1.16.0
text-unidecode==1.3
tzdata==2023.3
utility @ git+http://tanlab.ad.bcm.edu:3000/TanLab/utility.git@641edf2e7a83983484f60e7eaaf030c7c1fb4095
